Split lazyframe methods
